+++ static/css/style.css
@@ -1,0 +1,109 @@
+*{
+    margin: 0;
+    padding: 0;
+}
+
+.main{
+    width: 100%;
+    background: linear-gradient(to top,rgba(0,0,0,0.5)20%,rgba(0,0,0,0.5)20%), url('/static/image/home.jpg');
+    background-position: center;
+    background-size: cover;
+    height: 100vh;
+}
+
+.navbar{
+    width: 1200px;
+    height: 75px;
+    margin: auto;
+}
+.menu{
+    width: 400px;
+    float: left;
+    height: 70px;
+    color: red;
+}
+
+ul{
+    float: left;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+ul li{
+    float: left;
+    list-style: none;
+    margin-left: 80px ;
+    margin-top: 27px;
+    font-size: 35px;
+}
+
+ul li a{
+    text-decoration: none;
+    color: rgb(233, 237, 240);
+    font-family: Arial;
+    font-weight: bold;
+    transition: 1 sec ease-in-out;
+}
+
+ul li a:hover{
+    color: red;
+}
+
+.content{
+    width: 1200px;
+    height: auto;
+    margin: auto;
+    color: #fff;
+    position: relative;
+    margin-top: 25px;
+}
+
+.content .par{
+    padding-left: 20px;
+    padding-bottom: 25px;
+    font-family: arial;
+    letter-spacing: 1.2px;
+    line-height: 30px;
+}
+
+.content h1{
+    font-family: 'Times New Roman';
+    font-size: 50px;
+    padding-left: 20px;
+    margin-top: 9%;
+    letter-spacing: 2px;
+}
+
+.content .cn{
+    width: 160px;
+    height: 40px;
+    background: rgb(4, 237, 124);
+    border: none;
+    margin-bottom: 10px;
+    margin-left: 20px;
+    font-size: 18px;
+    border-radius: 10px;
+    cursor: pointer;
+    transition: 0.2s ease;
+}
+
+.content:hover{
+    color: rgb(1, 21, 249);
+    font-weight: bold;
+}
+
+.content .cn{
+    text-decoration: none;
+    color: #000;
+    transition: 1s ease;
+}
+
+.cn:hover{
+    background-color: #fff;
+
+}
+.content span{
+    color: chocolate;
+    font-size: 60px;
+}
