--- video_capture/app.py
@@ -1,46 +1,0 @@
-import mimetypes
-import cv2
-from flask import Flask,render_template, Response
-
-app=Flask(__name__)
-
-@app.route('/')
-def index():
-    return render_template('index.html')
-
-
-@app.route('/cap')
-def cap():
-    cam = cv2.VideoCapture(0)
-
-    cv2.namedWindow("test")
-
-    img_counter = 0
-
-    while True:
-        ret, frame = cam.read()
-        if not ret:
-            print("failed to grab frame")
-            break
-        cv2.imshow("test", frame)
-
-        k = cv2.waitKey(1)
-        if k%256 == 27:
-        # ESC pressed
-            print("Escape hit, closing...")
-            break
-        elif k%256 == 32:
-        # SPACE pressed
-            img_name = "opencv_frame_{}.png".format(img_counter)
-            cv2.imwrite(img_name, frame)
-            print("{} written!".format(img_name))
-            img_counter += 1
-
-    cam.release()
-
-    cv2.destroyAllWindows()
-
-
-#captures images from video by clicking space bar
-
-app.run(debug=True)
